requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
greenlet==3.1.1
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
